+++ openeogeotrellis/config/config.py
@@ -136,3 +136,50 @@ class GpsBackendConfig(OpenEoBackendConfig):
     s3_bucket_name: str = os.environ.get("SWIFT_BUCKET", "OpenEO-data")
 
     fuse_mount_batchjob_s3_bucket: bool = smart_bool(os.environ.get("FUSE_MOUNT_BATCHJOB_S3_BUCKET", False))
+
+    """
+    Reading strategy for load_collection and load_stac processes:
+    load_by_target_partition: first apply a partitioner, then simply read data for each partition
+    This avoids moving data around after reading, allowing the reading step to be followed immediately by next processes.
+    load_per_product: group by source product filename, then perform reads, then regroup data according to optimal partitioner for subsequent processes.
+    This strategy is faster when there is a high overhead/latency of opening products, for instance observed when reading jpeg2000 on S3. It does require
+    the data to be moved around in the cluster for subsequent processing.
+
+    The default can be overridden by feature_flags.
+    """
+    default_reading_strategy: str = "load_by_target_partition"
+
+    """
+    Controls the default number of threads for the batch job executors Java Virtual Machine.
+    These threads are used for additional parallelism, for instance when reading data, but may cause threading issues if
+    used in combination with not thread-safe libraries.
+    When reading with GDAL, a lower number of threads may be beneficial, as GDAL also performs its own threading.
+    """
+    default_executor_threads_jvm: int = 10
+
+    """
+    The default number of datasets that will be cached in batch jobs. A large cache will increase memory usage.
+    """
+    default_gdal_dataset_cache_size: int = 16
+
+    """
+    The default cache size in megabytes, which is used by GDAL. This setting applies to batch jobs.
+    """
+    default_gdal_cachemax: int = 150
+
+    """
+    The default maximum number of executors for batch jobs. A high number of executors can cause high costs, as allocated
+    executors may be idle for a relatively high amount of time.
+    We set this to 20 in kubernetes environments.
+    """
+    default_max_executors: int = 100
+
+    default_driver_memory: str = "2G"
+
+    default_driver_memoryOverhead: str = "3G"
+
+    default_executor_memory: str = "2G"
+
+    default_executor_memoryOverhead: str = "3G"
+
+    default_executor_cores: int = 2